Predictions/requirements.txt
pandas
xgboost
scikit-learn
supabase
python-dotenv
